fix:  Fix to allow move OFTs to wire a Solana peer
Prior to this fix, it was impossible to wire a Solana peer.  This change makes determining
the contractAddress more flexible, by reading it from hardhat-deploy deployments for EVM
peers and taking it directly from the configuration for Solana peers.  Supporting other
blockchains is future work and considered out of scope for this PR.

Signed-off-by: Ryan Goulding <goulding@layerzerolabs.org>